backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
python-multipart==0.0.12
anthropic>=0.40.0
openai==1.54.5
httpx==0.27.2
playwright==1.48.0
pillow==11.0.0
reportlab==4.2.5
celery==5.4.0
redis==5.2.0
python-dotenv==1.0.1
cryptography==44.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
pandas==2.2.3
openpyxl==3.1.5
passlib==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9
cachetools==5.5.0
tenacity==9.0.0
python-json-logger==2.0.7
